+++ src/main/java/org/example/fanzip/payment/service/PaymentRollbackService.java
@@ -1,6 +1,8 @@
 package org.example.fanzip.payment.service;
 
 import lombok.RequiredArgsConstructor;
+import org.example.fanzip.global.exception.BusinessException;
+import org.example.fanzip.global.exception.payment.PaymentErrorCode;
 import org.example.fanzip.payment.domain.Payments;
 import org.example.fanzip.payment.dto.PaymentResponseDto;
 import org.example.fanzip.payment.repository.PaymentRepository;
@@ -15,7 +17,7 @@ public class PaymentRollbackService {
     public PaymentResponseDto refundedPaymentById(Long paymentId){
         Payments payments = paymentRepository.findById(paymentId);
         if (payments == null) {
-            throw new IllegalArgumentException("해당 결제 정보를 찾을 수 없습니다. paymentId=" + paymentId);
+            throw new BusinessException(PaymentErrorCode.PAYMENT_NOT_FOUND);
         }
         payments.refund();
         paymentRepository.updateStatus(payments);
