+++ src/test/java/org/example/fanzip/payment/config/RootConfigTest.java
@@ -3,8 +3,8 @@ package org.example.fanzip.payment.config;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
-import org.example.fanzip.config.RootConfig;
-import org.example.fanzip.config.YamlPropertySourceFactory;
+import org.example.fanzip.global.config.RootConfig;
+import org.example.fanzip.global.config.YamlPropertySourceFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
