+++ src/test/java/org/example/fanzip/user/mapper/UserMapperTest.java
@@ -1,7 +1,8 @@
 package org.example.fanzip.user.mapper;
 
 import lombok.extern.slf4j.Slf4j;
-import org.example.fanzip.config.RootConfig;
+import org.example.fanzip.global.config.RootConfig;
+//import org.example.fanzip.user.dto.AdditionalInfoDTO;
 import org.example.fanzip.user.dto.UserDTO;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
